+++ NEWS.md
@@ -1,6 +1,7 @@
-dietr 1.1.5
+dietr 1.1.6
 ============
 
 ## Minor changes
-* Fixed issues with functionality caused by rfishbase updates
-* Updated vignette and documentation.
+* Fixed issues with functionality caused by rfishbase updates that caused package to be removed from CRAN
+* Updated documentation.
+* Added support for GII, MFI, and Windell Significance in CompositeIndices